Dataflow build additions (#253)
* [Build] add a deployment_package step

* [Build] typo fix in build examples

* [Build] fix deployment_package step

* [Analysis] ensure python integers in fpgadataflow analysis passes

* [Build] report generation and other minor improvements

* [Docs] update build_dataflow docs

* [Build] latency est. fix

* [Build] check if fold config is None

* [Build] add ooc synthesis step

* [Deps] update finn-base

* [Util] out_of_context_synth: remove remote, use launch_process_helper

* [Build] include all outputs in examples configs

* [Docs] update build flow docs

* [Deps] update finn-base

* [Util] bugfix in launch_process_helper call

* [Docker] use interactive mode for builds

* [Build] enable pdb debugging for builds

* [Refactor] move build functions to own submodule

* [Test] build_dataflow: fix expected files

* [Build] report estimated resource total

* [Infra] remove old eggs

* [HLSCustomOp] introduce get_op_counts

only implemented for MVAU and VVAU for now

* [HLSCustomOp] extend get_op_counts to include params too

* [Analysis] introduce op_and_param_counts pass

* [Build] generate op/param counts as part of estimates + add doc

* [HLSCustomOp] assert if ap_int_max_w is too large

* [StreamingFC] fix ap_int_max_w calculation

* [Build] minor fix in step_generate_estimate_reports
