+++ docs/finn/source_code/finn.rst
@@ -3,6 +3,8 @@ FINN API
 ********
 The FINN sources are divided into different modules. They are listed below.
 
+.. note:: **Some of these functions and modules are located in the `finn-base` repository.**
+
 Modules
 =======
 
@@ -10,6 +12,7 @@ Modules
    :maxdepth: 1
 
    finn.analysis
+   finn.builder
    finn.core
    finn.custom_op
    finn.transformation
