+++ docs/finn/source_code/finn.util.rst
@@ -13,15 +13,6 @@ finn.util.basic
    :undoc-members:
    :show-inheritance:
 
-finn.util.build\_dataflow
---------------------------
-
-.. automodule:: finn.util.build_dataflow
-  :members:
-  :undoc-members:
-  :show-inheritance:
-
-
 finn.util.data\_packing
 ------------------------------
