+++ src/finn/qnn-data/build_dataflow/dataflow_build_config.json
@@ -1,13 +1,16 @@
 {
   "output_dir": "output_tfc_w1a1_Pynq-Z1",
   "target_fps": 100000,
-  "mvau_width_max": 10000,
+  "mvau_wwidth_max": 10000,
   "synth_clk_period_ns": 10.0,
   "board": "Pynq-Z1",
   "shell_flow_type": "vivado_zynq",
   "generate_outputs": [
-    "pynq_driver",
+    "estimate_reports",
     "stitched_ip",
-    "bitfile"
+    "pynq_driver",
+    "out_of_context_synth",
+    "bitfile",
+    "deployment_package"
   ]
 }
