+++ tests/util/test_build_dataflow.py
@@ -30,7 +30,7 @@ import pytest
 import pkg_resources as pk
 from shutil import copytree
 from finn.util.basic import make_build_dir
-from finn.util.build_dataflow import build_dataflow_directory
+from finn.builder.build_dataflow import build_dataflow_directory
 import os
 
 
@@ -44,9 +44,19 @@ def test_build_dataflow_directory():
     build_dataflow_directory(target_dir)
     # check the generated files
     output_dir = target_dir + "/output_tfc_w1a1_Pynq-Z1"
+    assert os.path.isfile(output_dir + "/build_dataflow.log")
+    assert os.path.isfile(output_dir + "/time_per_step.json")
+    assert os.path.isfile(output_dir + "/final_hw_config.json")
     assert os.path.isfile(output_dir + "/stitched_ip/ip/component.xml")
     assert os.path.isfile(output_dir + "/driver/driver.py")
+    assert os.path.isfile(output_dir + "/report/estimate_layer_cycles.json")
+    assert os.path.isfile(output_dir + "/report/estimate_layer_resources.json")
+    assert os.path.isfile(
+        output_dir + "/report/estimate_layer_config_alternatives.json"
+    )
+    assert os.path.isfile(output_dir + "/report/estimate_network_performance.json")
+    assert os.path.isfile(output_dir + "/report/ooc_synth_and_timing.json")
     assert os.path.isfile(output_dir + "/bitfile/finn-accel.bit")
     assert os.path.isfile(output_dir + "/bitfile/finn-accel.hwh")
-    assert os.path.isfile(output_dir + "/bitfile/resources.xml")
-    assert os.path.isfile(output_dir + "/bitfile/timing.rpt")
+    assert os.path.isfile(output_dir + "/report/post_synth_resources.xml")
+    assert os.path.isfile(output_dir + "/report/post_route_timing.rpt")
